+++ frontend/src/shared/components/PageTemplate/UnauthenticatedPageTemplate.tsx
@@ -1,14 +1,7 @@
 import { Suspense } from 'react';
-import {
-  Navigate,
-  Outlet,
-  ScrollRestoration,
-  useLocation,
-} from 'react-router-dom';
+import { Outlet, ScrollRestoration } from 'react-router-dom';
 
 import { Box, CircularProgress, CssBaseline } from '@mui/material';
-
-import { useAuth } from '#shared/context/Auth.context';
 
 import { FlexRow } from '../FlexBox/FlexBox';
 import './UnauthenticatedPage.scss';
@@ -19,12 +12,8 @@ export function UnauthenticatedPageTemplate() {
   //   is loading in its children
   // - <Outlet> is a React-Router component which displays the children pages
   //   defined in the router.tsx
-  const { isAuthenticated } = useAuth();
-  const location = useLocation();
 
-  return isAuthenticated ? (
-    <Navigate to="/" state={{ from: location }} replace />
-  ) : (
+  return (
     <Box sx={{ position: 'absolute', top: 0, bottom: 0, left: 0, right: 0 }}>
       <Box className={'content'} style={{ position: 'fixed' }}>
         <Box className="blur"></Box>
